[feat] #70 mission 참여자 수 조회 및 업데이트 로직 수정

+++ src/main/java/org/letscareer/letscareer/domain/mission/dto/response/GetMissionDetailResponseDto.java
@@ -13,7 +13,7 @@ public record GetMissionDetailResponseDto(
         String title,
         MissionStatusType missionStatusType,
         Integer attendanceCount,
-        Integer lateAttendanceCount,
+        Long lateAttendanceCount,
         LocalDateTime startDate,
         LocalDateTime endDate
 ) {
